Restructure project
Add fisrt tests
