+++ src/api/controllers/flights_controller.py
@@ -1,9 +1,9 @@
 import uuid
 
-from database import db
-from models.flights_model import Flights
-from models.user_bookings_model import UserBookings
-from models.users_model import Users
+from api.database import db
+from api.models.flights_model import Flights
+from api.models.user_bookings_model import UserBookings
+from api.models.users_model import Users
 
 
 def create_flight(json_data):
