+++ src/api/models/flights_model.py
@@ -1,7 +1,7 @@
 import sqlalchemy
 from sqlalchemy.orm import relationship
 
-from database import db
+from api.database import db
 
 
 class Flights(db.Model):
