+++ src/api/models/user_bookings_model.py
@@ -2,7 +2,7 @@ import sqlalchemy
 from sqlalchemy import ForeignKey
 from sqlalchemy.orm import relationship
 
-from database import db
+from api.database import db
 
 
 class UserBookings(db.Model):
