+++ src/api/routes/crud_bookings_route.py
@@ -1,16 +1,14 @@
-import re
-
 from flask import Blueprint, request
 from flask_expects_json import expects_json
 from sqlalchemy.exc import IntegrityError
 
-from controllers.bookings_controller import *
-from controllers.flights_controller import get_flight_by_flight_number
-from controllers.users_controller import get_user_by_uuid
-from database import db
-from services.jwt_required_decorators import admin_required, admin_or_user_id_required, \
+from api.controllers.bookings_controller import *
+from api.controllers.flights_controller import get_flight_by_flight_number
+from api.controllers.users_controller import get_user_by_uuid
+from api.database import db
+from api.services.jwt_required_decorators import admin_required, admin_or_user_id_required, \
     require_admin_or_user_to_book_a_flight
-from utils import handle_integrity_error
+from api.utils import handle_integrity_error
 
 crud_bookings_bp = Blueprint("bookings", __name__)
