+++ src/api/seed.py
@@ -3,15 +3,13 @@ import uuid
 
 import flask_bcrypt
 from dotenv import load_dotenv
-from sqlalchemy import create_engine, text
+from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
-
 from sqlalchemy_utils import database_exists, create_database
 
-from database import db
 from models.flights_model import Flights
-from models.users_model import Users
 from models.user_bookings_model import UserBookings
+from models.users_model import Users
 
 load_dotenv()
