+++ src/api/templates/register.html
